web: Move all of the actual player logic out from RufflePlayer and into InnerPlayer
